add delete function and delete gamecard page
